Finished Timeline changed some fonts. Added a table working on construcing data for that
