+++ intern_portal/app/resources/HallofInterns/components/InternTable.tsx
@@ -1,0 +1,32 @@
+import React from "react";
+
+const InternTable = () => {
+  return (
+    <div>
+      <div className="overflow-x-auto">
+        <table className="table">
+          <thead>
+            <tr>
+              <th></th>
+              <th>Name</th>
+              <th>Current Where Abouts</th>
+              <th>Internship</th>
+              <th>intern currently? </th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr className="bg-base-200">
+              <th>NA</th>
+              <td>NA</td>
+              <td>NA</td>
+              <td>NA</td>
+              <td>NA</td>
+            </tr>
+          </tbody>
+        </table>
+      </div>
+    </div>
+  );
+};
+
+export default InternTable;
