+++ intern_portal/app/resources/Timeline/Components/Timeline.tsx
@@ -1,8 +1,13 @@
 import React from "react";
+import { Inter } from "next/font/google";
 
+const font = Inter({
+  subsets: ["latin"],
+  weight: ["400", "700"],
+});
 const Timeline = () => {
   return (
-    <div>
+    <div className={`font.className`}>
       <ul className="timeline timeline-snap-icon max-md:timeline-compact timeline-vertical">
         <li>
           <div className="timeline-middle">
@@ -22,7 +27,7 @@ const Timeline = () => {
           <div className="timeline-start mb-10 md:text-end">
             <time className="font-mono italic">June 3</time>
             <div className="text-lg font-black">Orientation</div>
-            This is your onboarding process not sure what else to say
+            This is your onboarding day. First day at work!
           </div>
           <hr />
         </li>
@@ -45,6 +50,10 @@ const Timeline = () => {
           <div className="timeline-end mb-10">
             <time className="font-mono italic">June 10</time>
             <div className="text-lg font-black">Capstone Meeting #1</div>
+            8:00am Business Communications for STEM Interns 9:00am Government
+            Contracting for STEM Interns 10:15am Huntsville interns will be in
+            Quandary conference room. 11:30am Lunch with STEM Interns 12:30am
+            Capstone Project
           </div>
           <hr />
         </li>
@@ -67,7 +76,10 @@ const Timeline = () => {
           <div className="timeline-start mb-10 md:text-end">
             <time className="font-mono italic">June 24</time>
             <div className="text-lg font-black">Capstone Meeting #2</div>
-            {/* Description tmr */}
+            8:00am Business Communications for STEM Interns 9:00am Government
+            Contracting for STEM Interns 10:15am Huntsville interns will be in
+            Quandary conference room. 11:30am Lunch with STEM Interns 12:30am
+            Capstone Project
           </div>
           <hr />
         </li>
@@ -88,57 +100,14 @@ const Timeline = () => {
             </svg>
           </div>
           <div className="timeline-end mb-10">
-            <time className="font-mono italic">June 8</time>
-            <div className="text-lg font-black">Capstone Meeting #3</div>
-            {/* Description tmr */}
-          </div>
-          <hr />
-        </li>
-        <li>
-          <hr />
-          <div className="timeline-middle">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 20 20"
-              fill="currentColor"
-              className="h-5 w-5"
-            >
-              <path
-                fillRule="evenodd"
-                d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.857-9.809a.75.75 0 00-1.214-.882l-3.483 4.79-1.88-1.88a.75.75 0 10-1.06 1.061l2.5 2.5a.75.75 0 001.137-.089l4-5.5z"
-                clipRule="evenodd"
-              />
-            </svg>
-          </div>
-          <div className="timeline-start mb-10 md:text-end">
-            <time className="font-mono italic">June 22</time>
-            <div className="text-lg font-black">Capstone Meeting #4</div>
-            {/* Description tmr */}
-          </div>
-        </li>
-        <li>
-          <hr />
-          <div className="timeline-middle">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 20 20"
-              fill="currentColor"
-              className="h-5 w-5"
-            >
-              <path
-                fillRule="evenodd"
-                d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.857-9.809a.75.75 0 00-1.214-.882l-3.483 4.79-1.88-1.88a.75.75 0 10-1.06 1.061l2.5 2.5a.75.75 0 001.137-.089l4-5.5z"
-                clipRule="evenodd"
-              />
-            </svg>
-          </div>
-          <div className="timeline-start mb-10 md:text-end">
             <time className="font-mono italic">July 8</time>
-            <div className="text-lg font-black">Capstone Meeting #5</div>
-            {/* Description tmr */}
+            <div className="text-lg font-black">Capstone Meeting #3</div>
+            8:00am Decision Making/Critical Thinking for STEM Interns 9:30am
+            Handling Conflcit 11:30am Lunch with STEM Interns 12:30am Capstone
+            Project
           </div>
+          <hr />
         </li>
-
         <li>
           <hr />
           <div className="timeline-middle">
@@ -157,11 +126,36 @@ const Timeline = () => {
           </div>
           <div className="timeline-start mb-10 md:text-end">
             <time className="font-mono italic">July 22</time>
-            <div className="text-lg font-black">Capstone Meeting #6</div>
-            {/* Description tmr */}
+            <div className="text-lg font-black">Capstone Meeting #4</div>
+            8:00am Personal and Career Development 9:30am Presentation Skills
+            11:00am Lunch with STEM Interns 12:30am Capstone Project
           </div>
+          <hr />
         </li>
-
+        <li>
+          <hr />
+          <div className="timeline-middle">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 20 20"
+              fill="currentColor"
+              className="h-5 w-5"
+            >
+              <path
+                fillRule="evenodd"
+                d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.857-9.809a.75.75 0 00-1.214-.882l-3.483 4.79-1.88-1.88a.75.75 0 10-1.06 1.061l2.5 2.5a.75.75 0 001.137-.089l4-5.5z"
+                clipRule="evenodd"
+              />
+            </svg>
+          </div>
+          <div className="timeline-end mb-10">
+            <time className="font-mono italic">August 5</time>
+            <div className="text-lg font-black">Capstone Meeting #5</div>
+            8:00am Capstone Intensive 11:00am Lunch With Stacey Springer 12:00pm
+            Capstone Intensive
+          </div>
+          <hr />
+        </li>
         <li>
           <hr />
           <div className="timeline-middle">
@@ -179,12 +173,12 @@ const Timeline = () => {
             </svg>
           </div>
           <div className="timeline-start mb-10 md:text-end">
-            <time className="font-mono italic">Aug 5</time>
-            <div className="text-lg font-black">Last Capstone meeting</div>
-            {/* Description tmr */}
+            <time className="font-mono italic">August 8</time>
+            <div className="text-lg font-black">Capstone Presentation!</div>
+            11:00am presenting capstone project!
           </div>
+          <hr />
         </li>
-
         <li>
           <hr />
           <div className="timeline-middle">
@@ -201,36 +195,12 @@ const Timeline = () => {
               />
             </svg>
           </div>
-          <div className="timeline-start mb-10 md:text-end">
-            <time className="font-mono italic">Aug 8</time>
-            <div className="text-lg font-black">Capstone Presentation</div>
-            {/* Description tmr */}
+          <div className="timeline-end mb-10">
+            <time className="font-mono italic">August 16</time>
+            <div className="text-lg font-black">Last day for Interns</div>
+            Last day of your internship!
           </div>
-        </li>
-
-        <li>
           <hr />
-          <div className="timeline-middle">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              viewBox="0 0 20 20"
-              fill="currentColor"
-              className="h-5 w-5"
-            >
-              <path
-                fillRule="evenodd"
-                d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.857-9.809a.75.75 0 00-1.214-.882l-3.483 4.79-1.88-1.88a.75.75 0 10-1.06 1.061l2.5 2.5a.75.75 0 001.137-.089l4-5.5z"
-                clipRule="evenodd"
-              />
-            </svg>
-          </div>
-          <div className="timeline-start mb-10 md:text-end">
-            <time className="font-mono italic">Aug 16</time>
-            <div className="text-lg font-black">
-              Last Day of your Internship
-            </div>
-            {/* Description tmr */}
-          </div>
         </li>
       </ul>
     </div>
